...G...PS. [ZBXNEXT-9338] returned a typecast
Unless we end up running on a 16-bit machine, we're expecting 32-bit
ints. For the sake of appeasing the %x conversion specifier, cast to the
expected type and hope compilers in the world where unsigned ints aren't
32-bit wide will warn us about the precision loss.

+++ src/libs/zbxregexp/zbxregexp.c
@@ -61,7 +61,7 @@ static char	*decode_pcre2_compile_error(int error_code, PCRE2_SIZE error_offset,
 	if (0 > (ret = pcre2_get_error_message(error_code, (PCRE2_UCHAR *)buf, sizeof(buf))))
 		return zbx_dsprintf(NULL, "pcre2_get_error_message(%d, ...) failed with error %d", error_code, ret);
 
-	return zbx_dsprintf(NULL, "%s, position %zu, flags:0x%x", buf, (size_t)error_offset, flags);
+	return zbx_dsprintf(NULL, "%s, position %zu, flags:0x%x", buf, (size_t)error_offset, (unsigned int)flags);
 #undef BUF_SIZE
 }
